Update CanMessage to use uint8_t for realistic byte-level CAN data

+++ include/CanMessage.hpp
@@ -3,16 +3,16 @@
 
 #include <string>     // For using std::string
 #include <vector>     // For using std::vector
-
+#include <cstdint>
 // represents a fake CAN message
 class CanMessage {
 public:
     int id;                       // CAN message ID (e.g. 0x120)
     int dlc;                      // Data Length Code: how many data bytes
-    std::vector<int> data;        // Data payload (up to 8 values in real CAN)
+    std::vector<uint8_t> data;         //  byte-level data
 
     // Constructor: creates a new CAN message
-    CanMessage(int input_id, int input_dlc, std::vector<int> input_data) {
+    CanMessage(int input_id, int input_dlc, std::vector<uint8_t> input_data) {
         id = input_id; // 0x120 canid
         dlc = input_dlc; // number of byte 
         data = input_data;
